prometheus-libvirt-exporter/requirements.txt
libvirt-python
prometheus_client